Project-Blog/blog/requirements.txt
asgiref==3.7.2
Django==4.2.7
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
django-ckeditor==6.7.0
django-js-asset==2.2.0